+++ lib/src/variant/variants/fairy.dart
@@ -2,10 +2,25 @@ part of '../variant.dart';
 
 /// Variants where the focus is on novel piece types.
 class FairyVariants {
-  static Variant hoppelPoppel() => Variant.standard()
-      .copyWith(name: 'Hoppel-Poppel')
-      .withPieces({'N': PieceType.knibis(), 'B': PieceType.biskni()});
+  /// Knights capture like bishops, bishops capture like knights.
+  static Variant hoppelPoppel() => Variant.standard().withPieces({
+        'N': PieceType.knibis(),
+        'B': PieceType.biskni(),
+      }).copyWith(
+        name: 'Hoppel-Poppel',
+        materialConditions: MaterialConditions.none,
+      );
 
+  /// Rooks capture like knights, knights capture like rooks.
+  static Variant newZealand() => Variant.standard().withPieces({
+        'R': PieceType.rookni(),
+        'N': PieceType.kniroo(),
+      }).copyWith(
+        name: 'New Zealand Chess',
+        materialConditions: MaterialConditions.none,
+      );
+
+  /// https://en.wikipedia.org/wiki/Grasshopper_chess
   static Variant grasshopper() =>
       Variant.standard().withPiece('G', PieceType.grasshopper()).copyWith(
             name: 'Grasshopper Chess',
@@ -15,6 +30,15 @@ class FairyVariants {
             materialConditions: MaterialConditions.none,
           );
 
+  /// Knights are replaced with nightriders.
+  static Variant nightrider() => Variant.standard()
+      .copyWith(
+        name: 'Nightrider Chess',
+        materialConditions: MaterialConditions.none,
+      )
+      .withPiece('N', PieceType.nightrider());
+
+  /// https://en.wikipedia.org/wiki/Berolina_pawn#Berolina_chess
   static Variant berolina() => Variant.standard()
       .withPiece('P', PieceType.berolinaPawn())
       .copyWith(name: 'Berolina Chess');